+++ vsreact/js/src/components.ts
@@ -69,6 +69,23 @@ export type { PianoKeyboardProps } from "./keyboard";
 export { StepSequencer } from "./sequencer";
 export type { StepSequencerProps } from "./sequencer";
 export {
+  ADSREnvelope,
+  ParamADSREnvelope,
+  PitchBend,
+  ModWheel,
+  ParamModWheel,
+  ParamPitchBend,
+} from "./synth";
+export type {
+  ADSRKey,
+  ADSREnvelopeProps,
+  ParamADSREnvelopeProps,
+  PitchBendProps,
+  ModWheelProps,
+  ParamModWheelProps,
+  ParamPitchBendProps,
+} from "./synth";
+export {
   MacroPad,
   ParamMacroPad,
   HardwareKnob,